Logging Touchups
Removed type from the Module class & from the logging tab.

Altered the logging click handler so it adds the module to the programs
stored modules (the "landing grid").

Added some input validation for code, xc, and yc.

Altered the way the storetable is populated.

Removed the placeholder html from the two tabs with actual content.